+++ packages/rax-textinput/src/index.tsx
@@ -5,7 +5,8 @@ import {
   createElement,
   ForwardRefExoticComponent,
   useEffect,
-  useState
+  useState,
+  useMemo
 } from 'rax';
 import { isWeex, isWeb, isNode, isMiniApp } from 'universal-env';
 import setNativeProps from 'rax-set-native-props';
@@ -45,6 +46,16 @@ function genEventObject(event): EventObject {
   };
 }
 
+/**
+ * Judge whether the val is truthy or zero value
+ * Because number 0 should be recognized as valid input too
+ * @param val
+ * @returns
+ */
+function isTruthyOrZero(val) {
+  return val === 0 ? true : !!val;
+}
+
 const TextInput: ForwardRefExoticComponent<TextInputProps> = forwardRef(
   (props, ref) => {
     const refEl = useRef<TextInputElement>(null);
@@ -76,7 +87,7 @@ const TextInput: ForwardRefExoticComponent<TextInputProps> = forwardRef(
       secureTextEntry,
       style,
       placeholderColor = '#999999',
-      value,
+      value: propsValue,
       defaultValue,
       controlled
     } = props;
@@ -86,6 +97,13 @@ const TextInput: ForwardRefExoticComponent<TextInputProps> = forwardRef(
         : typeof keyboardTypeMap[keyboardType] === 'undefined'
           ? keyboardType
           : keyboardTypeMap[keyboardType];
+
+    let value = isTruthyOrZero(propsValue) ? propsValue : '';
+
+    useMemo(() => {
+      value = isTruthyOrZero(propsValue) ? propsValue :
+        isTruthyOrZero(defaultValue) ? defaultValue : '';
+    }, []);
 
     // Check is type supported or not
     // Use isWeb to exclude web-view
@@ -129,7 +147,7 @@ const TextInput: ForwardRefExoticComponent<TextInputProps> = forwardRef(
     };
 
     // Diff with web readonly attr, `disabled` must be boolean value
-    const disbaled = Boolean(editable !== undefined && !editable);
+    const disabled = Boolean(editable !== undefined && !editable);
     const rows = numberOfLines || maxNumberOfLines;
     useImperativeHandle(ref, () => {
       return {
@@ -172,9 +190,9 @@ const TextInput: ForwardRefExoticComponent<TextInputProps> = forwardRef(
             }}
             row={rows}
             rows={rows}
-            disabled={disbaled}
+            disabled={disabled}
             onChange={handleChange}
-            value={value || defaultValue}
+            value={value}
             confirm-type={confirmType}
             show-count={showCount}
             onInput={(e: Rax.ChangeEvent<HTMLTextAreaElement>) => {
@@ -204,8 +222,8 @@ const TextInput: ForwardRefExoticComponent<TextInputProps> = forwardRef(
               placeholderColor
             }}
             type={type}
-            disabled={disbaled}
-            value={value || defaultValue}
+            disabled={disabled}
+            value={value}
             confirm-type={confirmType}
             random-Number={randomNumber}
             selection-start={selectionStart}